remove _common from tools, part 1
