--- tests/core/tools/dummy/dummy.py
@@ -1,6 +1,0 @@
-def setup(chip):
-    chip.set('tool', 'dummy', 'exe', 'echo')
-
-
-def parse_version(stdout):
-    raise IndexError('This is an index error')
