--- tests/core/tools/dummy/import.py
@@ -1,7 +1,0 @@
-from core.tools.dummy.dummy import setup as dummy_setup
-
-
-def setup(chip):
-    dummy_setup(chip)
-    chip.add('tool', 'dummy', 'task', 'import', 'output', chip.top() + '.v',
-             step='import', index='0')
