--- tests/core/tools/fake/bar.py
@@ -1,2 +1,0 @@
-def run():
-    return 0
