--- tests/core/tools/fake/baz.py
@@ -1,2 +1,0 @@
-def run():
-    return 0
