--- tests/core/tools/fake/foo.py
@@ -1,2 +1,0 @@
-def run():
-    return 0
