--- tests/core/tools/run/run.py
@@ -1,12 +1,0 @@
-def setup(chip):
-    tool = 'run'
-
-    chip.set('tool', tool, 'exe', 'sh')
-
-
-def parse_version(stdout):
-    '''
-    Version check based on stdout
-    Depends on tool reported string
-    '''
-    return '0'
