+++ tests/scheduler/tools/echo.py
@@ -1,0 +1,12 @@
+from siliconcompiler import TaskSchema
+
+
+class EchoTask(TaskSchema):
+    def __init__(self):
+        super().__init__()
+
+    def tool(self):
+        return "echo"
+
+    def task(self):
+        return "echo"
