+++ tests/tools/inputimporter/__init__.py
@@ -1,31 +1,34 @@
-'''
-This tool is used to copy files into the SiliconCompiler flow. It copies files
-into the output directory to be absorbed into downstream nodes.
-'''
+import shutil
 
-from siliconcompiler.tools._common import get_tool_task
+import os.path
 
-
-################################
-# Setup Tool (pre executable)
-################################
-def setup(chip):
-    tool = 'inputimporter'
-    step = chip.get('arg', 'step')
-    index = chip.get('arg', 'index')
-
-    _, task = get_tool_task(chip, step, index)
-
-    chip.set('tool', tool, 'exe', 'cp')
-
-    chip.set('tool', tool, 'task', task, 'regex', 'warnings', r'^\[WRN:',
-             step=step, index=index, clobber=False)
-    chip.set('tool', tool, 'task', task, 'regex', 'errors', r'^\[(ERR|FTL|SNT):',
-             step=step, index=index, clobber=False)
+from siliconcompiler import TaskSchema
 
 
-################################
-# Version Check
-################################
-def parse_version(stdout):
-    raise IndexError('This is an index error')
+class ImporterTask(TaskSchema):
+    '''
+    Import (copy) files into the output folder.
+    '''
+    def __init__(self):
+        super().__init__()
+
+        self.add_parameter("input_files", "[file]", "input files to copy")
+
+    def tool(self):
+        return "testing"
+
+    def task(self):
+        return "importer"
+
+    def setup(self):
+        super().setup()
+
+        self.add_required_tool_key("var", "input_files")
+
+        for file in self.get("var", "input_files"):
+            self.add_output_file(os.path.basename(file))
+
+    def run(self):
+        for file in self.find_files("var", "input_files"):
+            shutil.copy2(file, "outputs/")
+        return 0
